requirements.txt
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fire==0.6.0
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
idna==3.6
importlib_resources==6.4.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.3
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==5.26.0
pyarrow==15.0.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
ray==2.10.0
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sympy==1.12
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.4.0
torch==2.2.1+cu121
torch-summary==1.4.5
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.18.1
